packages/python/adapter/examples/basic/requirements.txt
otlp-stdout-adapter>=0.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
requests 